sennder_task/requirements.txt
Django==3.2.1
mysqlclient==1.4.6
djangorestframework==3.12.2
celery==5.1.0
redis==3.5.3
requests==2.25.1
drf-yasg==1.20.0